third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-justify-content-expected.txt

This test checks that the justify-content property is applied correctly when using content-position values.

direction: LTR | justify-content: 'center'

PASS
direction: LTR | justify-content: 'left'

PASS
direction: LTR | justify-content: 'right'

PASS
direction: LTR | justify-content: 'start'

PASS
direction: LTR | justify-content: 'end'

PASS
direction: LTR | justify-content: 'flex-start'

PASS
direction: LTR | justify-content: 'flex-end'

PASS
direction: LTR | justify-content: 'space-between' (fallback to 'start')

PASS
direction: LTR | justify-content: 'space-around' (fallback to 'center')

PASS
direction: LTR | justify-content: 'space-evenly' (fallback to 'center')

PASS
direction: LTR | justify-content: 'stretch' (fallback to start)

PASS
direction: LTR | justify-content: space-between left (fallback to left)

PASS
direction: LTR | justify-content: 'space-around end' (fallback to end)

PASS
direction: LTR | justify-content: 'space-evenly right' (fallback to 'right')

PASS
direction: LTR | justify-content: 'stretch end' (fallback to 'end')

PASS
direction: LTR | justify-content: 'auto' (resolved to 'start')

PASS
direction: RTL | justify-content: 'center'

PASS
direction: RTL | justify-content: 'left'

PASS
direction: RTL | justify-content: 'right'

PASS
direction: RTL | justify-content: 'start'

PASS
direction: RTL | justify-content: 'end'

PASS
